ia-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
google-genai
pytest
